refactor; I think this is ready for testing
