--- deploy.nu
@@ -1,5 +1,0 @@
-#cargo build --release
-#cp target\i686-pc-windows-msvc\release\amax_pfp.dll D:\Blur\amax\dlls\amax_pfp.dll
-
-cargo build
-cp target\i686-pc-windows-msvc\debug\amax_pfp.dll D:\Blur\amax\dlls\amax_pfp.dll
