+++ src/data/mod.rs
@@ -1,0 +1,2 @@
+pub mod gamer_picture_manager;
+pub mod net_racer;
