--- x.cmd
@@ -1,1 +1,0 @@
-nu deploy.nu
